% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{backward.crbm}
\alias{backward.crbm}
\title{Reconstruction using a Conditional Restricted Boltzmann Machine Function}
\usage{
backward.crbm(crbm, act.input, history)
}
\arguments{
\item{crbm}{A trained CRBM using train.crbm() function.}

\item{act.input}{A matrix with data, one activation example per row.}

\item{history}{A matrix with the visible history data, one example per row.
This matrix must be the size of (newdata - 1 + delay) examples. Sample
i from activations uses history (i - 1 - delay) to (i - 1) as visible
history. The selected slice will contain OLDER values first (t-delay to t-1)
\preformatted{
           act.input        history

          <--hidden-->   <--visible-->
           __________ 
[n,]    A |__________| _  ___________                    t=d+n
        | |__________|   |___________| A        [d+n-1,] t=d+n-1
   n    | |__________|   |___________| |                 .
samples | |__________|   |___________| |  n - 1          .
        | |__________|   |___________| | samples         .
        | |__________|   |___________| |                 .
[1,]    V |__________| _ |___________| V        [d+1,]   t=d+1
                         |___________| A        [d,]     t=d
                         |___________| |                 .
                         |___________| |                 .
                         |___________| |  delay          .
                         |___________| |                 .
                         |___________| V        [1,]     t=1
}}
}
\description{
Function to reconstruct inputs from a CRBM (Forward pass). Returns a
prediction matrix.
}
\examples{
test_X <- t(array(c(1, 1, 0, 0, 0, 0,
                    0, 1, 1, 1, 0, 0,
                    0, 1, 0, 1, 0, 0,
                    0, 1, 1, 1, 1, 0,
                    0, 0, 0, 1, 1, 0), c(6,5)));
activations <- forward.crbm(crbm1, test_X);

valid_activations <- activations[(crbm$delay+1):nrow(activations)];

i <- 3;
single_act <- valid_activations[i,,drop=FALSE];
corresp_history <- test_X[(i - 1 - crbm$delay):(i - 1),];
reconstruct <- backward.crbm(crbm1, single_act, corresp_history);

i <- 3;
some_activations <- valid_activations[i:(i+2),];
corresp_history <- test_X[(i - 1 - crbm$delay):(i - 1 - crbm$delay + 2),];
reconstruct <- backward.crbm(crbm1, some_activations, corresp_history);

## Motion (fragment) Example
data(motionfrag)

crbm_mfrag <- train.crbm(motionfrag$batchdata, motionfrag$seqlen,
                         batch_size = 100, n_hidden = 100, delay = 6,
                         training_epochs = 200, learning_rate = 1e-3,
                         momentum = 0.5, rand_seed = 1234);

activations <- forward.crbm(crbm_mfrag, motionfrag$batchdata);

valid_activations <- activations[crbm_mfrag$delay:nrow(activations)];
corresp_history <- motionfrag$batchdata[1:(nrow(motionfrag$batchdata)-1),];
reconstruct <- backward.crbm(crbm_mfrag, valid_activations, corresp_history);
}
\keyword{CRBM,}
\keyword{RBM}
