% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{forecast.crbm}
\alias{forecast.crbm}
\title{Predicting using a Conditional Restricted Boltzmann Machine Function}
\usage{
forecast.crbm(crbm, sequence, n_samples = 1, n_gibbs = 30)
}
\arguments{
\item{crbm}{A trained CRBM using train.crbm() function.}

\item{sequence}{Sequence for first input and its history. A matrix with data,
one example per row. Must contain more rows than crbm delay.}

\item{n_samples}{Number of samples to generate forward. Default = 1.}

\item{n_gibbs}{Number of alternating Gibbs steps per iteration. Default = 30-}
}
\description{
Function to, given initialization of visibles and matching history, generate
n_samples in future.
}
\examples{
train_X <- t(array(c(1, 1, 1, 0, 0, 0, # Sequence 1
                     1, 0, 1, 0, 0, 0,
                     1, 1, 1, 0, 0, 0,
                     0, 0, 0, 1, 1, 0,
                     0, 0, 1, 0, 1, 0,
                     1, 0, 1, 0, 0, 0,
                     1, 0, 1, 0, 1, 0, # Sequence 2
                     0, 0, 0, 1, 0, 0,
                     0, 0, 0, 0, 1, 0,
                     0, 0, 0, 1, 1, 0,
                     0, 1, 1, 0, 1, 0,
                     1, 0, 1, 1, 1, 0), c(12, 6)));
crbm1 <- train.crbm(train_X, seqlen = c(6, 6), delay = 2);

## Once trained, predict sequence
data_X <- t(array(c(1, 1, 0, 0, 0, 0,
                    0, 1, 1, 1, 0, 1,
                    0, 1, 0, 1, 0, 0,
                    0, 1, 0, 1, 1, 1,
                    1, 1, 0, 1, 1, 0,
                    1, 1, 0, 0, 0, 0,
                    1, 1, 1, 1, 0, 0,
                    0, 0, 0, 1, 1, 0), c(6,8)));
res3 <- forecast.crbm(crbm1, data_X[1:2, ], 5, 30);

## Motion (fragment) Example
data(motionfrag)

crbm_mfrag <- train.crbm(motionfrag$batchdata, motionfrag$seqlen,
                         batch_size = 100, n_hidden = 100, delay = 6,
                         training_epochs = 200, learning_rate = 1e-3,
                         momentum = 0.5, rand_seed = 1234);

offset <- crbm_mfrag$delay + 1;
fc <- forecast.crbm(crbm_mfrag, motionfrag$batchdata[1:offset,], 50, 30);
}
\keyword{CRBM,}
\keyword{FORECAST}
\keyword{RBM,}
