% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{forward.crbm}
\alias{forward.crbm}
\title{Feature generation using a Conditional Restricted Boltzmann Machine Function}
\usage{
forward.crbm(crbm, newdata)
}
\arguments{
\item{crbm}{A trained CRBM using train.crbm() function.}

\item{newdata}{A matrix with data, one example per row. Must contain
more rows than crbm delay.
\preformatted{
          <--visible-->
           ___________ 
[n,]    A |___________|                            t=n
        | |___________|                            .
        | |___________|                            .
        | |___________|                            .
   n    | |___________|                            .
samples | |___________|                      A     .
        | |___________|                      |     .
        | |___________|                      |     .
        | |___________| ---->  Input  --+    |     t=d+1
[d,]  A | |___________| --+             | Sliding  t=d
      | | |___________|   |             | window   .
delay | | |___________|   |-> History --+          .
      | | |___________|   |                        .
[1,]  V V |___________| --+                        t=1
}}
}
\description{
Function to Generate features from a CRBM (Forward pass). Returns a
prediction matrix.
}
\examples{
test_X <- t(array(c(1, 1, 0, 0, 0, 0,
                    0, 1, 1, 1, 0, 0,
                    0, 0, 0, 1, 1, 0), c(6,3)));
activations <- forward.crbm(crbm1, test_X);

## Motion (fragment) Example
data(motionfrag)

crbm_mfrag <- train.crbm(motionfrag$batchdata, motionfrag$seqlen,
                         batch_size = 100, n_hidden = 100, delay = 6,
                         training_epochs = 200, learning_rate = 1e-3,
                         momentum = 0.5, rand_seed = 1234);

activations <- forward.crbm(crbm_mfrag, motionfrag$batchdata);
}
\keyword{CRBM,}
\keyword{RBM}
