% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{forward.rbm}
\alias{forward.rbm}
\title{Featurizing using a Restricted Boltzmann Machine Function}
\usage{
forward.rbm(rbm, newdata)
}
\arguments{
\item{rbm}{A trained RBM using train.rbm() function.}

\item{act.input}{A matrix with activation data, one example per row.}
}
\description{
Function to Activate feaures from a RBM (Forward pass). Returns an
activation matrix.
}
\examples{
test_X <- t(array(c(1, 1, 0, 0, 0, 0,
                    0, 0, 0, 1, 1, 0), c(6,2)));
res <- predict(rbm1, test_X);
reconstruction <- backward.rbm(rbm1, res$activation);

activation <- forward.rbm(rbm1, test_X);
reconstruction <- backward.rbm(rbm1, activation);

# The MNIST example
data(mnist)

training.num <- data.matrix(mnist$train$x)/255;
rbm_mnist <- train.rbm(n_hidden = 30, dataset = training.num,
                       learning_rate = 1e-3, training_epochs = 10,
                       batch_size = 10, momentum = 0.5);

passed_data <- predict(rbm_mnist, training.num);
reconstruction <- backward.rbm(rbm_mnist, passed_data$activation);

activation <- forward.rbm(rbm_mnist, training.num);
reconstruction <- backward.rbm(rbm_mnist, activation);
}
\keyword{RBM}
