% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{train.crbm}
\alias{train.crbm}
\title{Training a Conditional Restricted Boltzmann Machine Function}
\usage{
train.crbm(dataset, seqlen, batch_size = 1, n_hidden = 3, delay = 6,
  training_epochs = 1000, learning_rate = 0.01, momentum = 0.8,
  rand_seed = 1234)
}
\arguments{
\item{dataset}{A matrix with data, one example per row.}

\item{seqlen}{A vector with the lenght of sequences in dataset.}

\item{batch_size}{Number of examples per training mini-batch. Default = 1.}

\item{n_hidden}{Number of hidden units in the CRBM. Default = 3.}

\item{delay}{Size of delay window for a sequence. Default = 6.}

\item{training_epochs}{Number of training epochs. Default = 1000.}

\item{learning_rate}{The learning rate for training. Default = 0.01.}

\item{momentum}{The momentum for training. Default = 0.8.}

\item{rand_seed}{Random seed. Default = 1234.}
}
\description{
This function trains a CRBM. Returns a CRBM in list form
}
\examples{
train_X <- t(array(c(1, 1, 1, 0, 0, 0, # Sequence 1
                     1, 0, 1, 0, 0, 0,
                     1, 1, 1, 0, 0, 0,
                     0, 0, 0, 1, 1, 0,
                     0, 0, 1, 0, 1, 0,
                     1, 0, 1, 0, 0, 0,
                     1, 0, 1, 0, 1, 0, # Sequence 2
                     0, 0, 0, 1, 0, 0,
                     0, 0, 0, 0, 1, 0,
                     0, 0, 0, 1, 1, 0,
                     0, 1, 1, 0, 1, 0,
                     1, 0, 1, 1, 1, 0), c(12, 6)));
crbm1 <- train.crbm(train_X, seqlen = c(6, 6), delay = 2);

## Motion (fragment) Example
data(motionfrag)

crbm_mfrag <- train.crbm(motionfrag$batchdata, motionfrag$seqlen,
                         batch_size = 100, n_hidden = 100, delay = 6,
                         training_epochs = 200, learning_rate = 1e-3,
                         momentum = 0.5, rand_seed = 1234);
}
\keyword{CRBM,}
\keyword{RBM}
